add InvalidAccess and Loading page
