+++ src/router.jsx
@@ -8,6 +8,7 @@ import MiniQuiz from '@/pages/miniquiz/MiniQuiz';
 import WorldCup from '@/pages/worldCup/WorldCupMain';
 import WorldCupResult from '@/pages/worldCup/WorldCupResult';
 import MiniQuizResult from '@/pages/miniquiz/MiniQuizResult';
+import InvalidAccess from '@/pages/miniquiz/InvalidAccess';
 
 const router = createBrowserRouter([
   {
@@ -33,6 +34,10 @@ const router = createBrowserRouter([
           {
             path: 'miniQuiz',
             element: <MiniQuiz />,
+          },
+          {
+            path: 'InvalidAccess',
+            element: <InvalidAccess />,
           },
           {
             path: 'miniQuizResult',
